requirements.txt
firebase-admin
livekit-api==1.0.7
livekit-agents==1.2.17